move: 패키지 이동
- 공통된 특성을 가지는 클래스끼리 같은 패키지로 묶기
