+++ eeos/src/main/java/com/blackcompany/eeos/auth/application/repository/OauthVerificationStorage.java
@@ -1,7 +1,7 @@
 package com.blackcompany.eeos.auth.application.repository;
 
 import com.blackcompany.eeos.auth.application.domain.OauthMemberModel;
-import com.blackcompany.eeos.auth.persistence.OAuthInfo;
+import com.blackcompany.eeos.auth.persistence.oauth.OAuthInfo;
 import java.util.UUID;
 
 public interface OauthVerificationStorage {
