+++ eeos/src/main/java/com/blackcompany/eeos/auth/application/service/DeactivateMemberService.java
@@ -5,7 +5,7 @@ import com.blackcompany.eeos.auth.application.event.DeletedMemberEvent;
 import com.blackcompany.eeos.auth.application.repository.OAuthMemberRepository;
 import com.blackcompany.eeos.auth.application.usecase.LogOutUsecase;
 import com.blackcompany.eeos.auth.application.usecase.WithDrawUsecase;
-import com.blackcompany.eeos.auth.persistence.AccountJpaRepository;
+import com.blackcompany.eeos.auth.persistence.account.AccountJpaRepository;
 import com.blackcompany.eeos.auth.persistence.InvalidTokenRepository;
 import com.blackcompany.eeos.member.application.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
