+++ eeos/src/main/java/com/blackcompany/eeos/auth/persistence/account/AccountEntity.java
@@ -1,4 +1,4 @@
-package com.blackcompany.eeos.auth.persistence;
+package com.blackcompany.eeos.auth.persistence.account;
 
 import com.blackcompany.eeos.common.persistence.BaseEntity;
 import jakarta.persistence.Column;
@@ -15,26 +15,30 @@ import lombok.NoArgsConstructor;
 import lombok.ToString;
 import lombok.experimental.SuperBuilder;
 
-@Entity
+@Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @ToString
 @SuperBuilder(toBuilder = true)
+@Entity
 @Table(
-		name = AuthorityEntity.NAME,
-		indexes = {@Index(name = "idx_member_id", columnList = AuthorityEntity.NAME + "_member_id")})
-@Getter
-public class AuthorityEntity extends BaseEntity {
+		name = AccountEntity.ENTITY_PREFIX,
+		indexes = {@Index(name = "idx_member_id", columnList = "account_member_id")})
+public class AccountEntity extends BaseEntity {
 
-	public static final String NAME = "authority";
+	public static final String ENTITY_PREFIX = "account";
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	@Column(name = ENTITY_PREFIX + "_id", nullable = false)
 	private Long id;
 
-	@Column(nullable = false, name = NAME + "_member_id")
+	@Column(name = ENTITY_PREFIX + "_member_id", nullable = false, unique = true)
 	private Long memberId;
 
-	@Column(nullable = false, name = NAME + "_role")
-	private String role;
+	@Column(name = ENTITY_PREFIX + "_login_id", nullable = false, unique = true)
+	private String loginId;
+
+	@Column(name = ENTITY_PREFIX + "_login_passwd", nullable = false)
+	private String passWd;
 }
