+++ eeos/src/main/java/com/blackcompany/eeos/auth/persistence/account/AccountJpaRepository.java
@@ -1,4 +1,4 @@
-package com.blackcompany.eeos.auth.persistence;
+package com.blackcompany.eeos.auth.persistence.account;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
