+++ eeos/src/main/java/com/blackcompany/eeos/auth/persistence/oauth/JpaOAuthMemberRepository.java
@@ -1,4 +1,4 @@
-package com.blackcompany.eeos.auth.persistence;
+package com.blackcompany.eeos.auth.persistence.oauth;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
