+++ eeos/src/main/java/com/blackcompany/eeos/member/application/service/CreateAdminMemberService.java
@@ -6,8 +6,8 @@ import com.blackcompany.eeos.auth.application.model.AuthorityModel;
 import com.blackcompany.eeos.auth.application.model.Role;
 import com.blackcompany.eeos.auth.application.repository.AuthorityRepository;
 import com.blackcompany.eeos.auth.application.repository.OAuthMemberRepository;
-import com.blackcompany.eeos.auth.persistence.AccountEntity;
-import com.blackcompany.eeos.auth.persistence.AccountJpaRepository;
+import com.blackcompany.eeos.auth.persistence.account.AccountEntity;
+import com.blackcompany.eeos.auth.persistence.account.AccountJpaRepository;
 import com.blackcompany.eeos.member.application.model.AdminInfo;
 import com.blackcompany.eeos.member.application.model.MemberModel;
 import com.blackcompany.eeos.member.application.repository.MemberRepository;
